pyproject.toml
[tool.poetry]
name = "marketing-ai"
version = "0.1.0"
description = "AI-Powered Competitor Ad Analysis & Content Recommendation System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
supabase = "^2.3.4"
httpx = "^0.26.0"
celery = {extras = ["redis"], version = "^5.3.6"}
openai = "^1.10.0"
google-generativeai = "^0.3.2"
anthropic = "^0.18.0"
pypdf = "^4.0.1"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
